Split AbstractProxyManager
Splits AbstractProxyManager into AbstractProxyManager and AbstractProvidingProxyManager. This allows a manager to use AbstractProxyManager without being able to provide new named beans.
